Clarify ground truthing module names after file split.
Rename base→frame_base, setup_frames→welcome_frame,
session_frames→annotation_frames, and plotting→annotation_plot
so each module name reflects its responsibility.
