+++ nps_active_space/ground_truthing/frame_base.py
